Split syscall.Exec into platform-specific files for Windows support
Move attachToTmuxSession, execAgent, execRuntime from helpers.go into
helpers_unix.go (syscall.Exec) and helpers_windows.go (os/exec.Command
with stdio passthrough and exit code propagation).

Move execBdShow from show.go into show_unix.go and show_windows.go
with the same pattern.

Fixes: gt-03t, gt-21i, gt-ko7


Executed-By: mayor
