+++ internal/cmd/show.go
@@ -2,10 +2,7 @@ package cmd
 
 import (
 	"fmt"
-	"os"
-	"os/exec"
 	"strings"
-	"syscall"
 
 	"github.com/spf13/cobra"
 )
@@ -45,38 +42,6 @@ func runShow(cmd *cobra.Command, args []string) error {
 	}
 
 	return execBdShow(args)
-}
-
-// execBdShow replaces the current process with 'bd show'.
-// Resolves the correct rig directory from the bead's prefix via routes.jsonl
-// so that rig-prefixed beads (e.g., myproject-abc) are found in their rig
-// database rather than only the town-level hq database. (GH#2126)
-func execBdShow(args []string) error {
-	bdPath, err := exec.LookPath("bd")
-	if err != nil {
-		return fmt.Errorf("bd not found in PATH: %w", err)
-	}
-
-	// Resolve the rig directory for the bead's prefix so bd runs from the
-	// correct working directory. Without this, bd may query the wrong database
-	// when inherited BEADS_DIR is set or when bd's routing doesn't handle
-	// cross-rig lookups from the town root.
-	if beadID := extractBeadIDFromArgs(args); beadID != "" {
-		if dir := resolveBeadDir(beadID); dir != "" && dir != "." {
-			_ = os.Chdir(dir)
-		}
-	}
-
-	// Strip BEADS_DIR from the environment so bd discovers the database from
-	// its working directory rather than using an inherited value that may point
-	// to the wrong (e.g., town-level) database.
-	env := stripEnvKey(os.Environ(), "BEADS_DIR")
-
-	// Build args: bd show <all-args>
-	// argv[0] must be the program name for exec
-	fullArgs := append([]string{"bd", "show"}, args...)
-
-	return syscall.Exec(bdPath, fullArgs, env)
 }
 
 // extractBeadIDFromArgs returns the first non-flag argument, which is the bead ID.
